Add Turk support to Rails server
I added a new entry point, /tasks/turk/<condition id>/, which can be
used for Turk workers to create a task. It turns out this wasn't
possible to do with a JavaScript-only frontend because the
Turk parameters (in particular turkSubmitTo) need to be available
to templates during the final step in order to construct a valid
submit button to submit the HIT for payment.

This commit stores all the Turk-supplied parameters as part of the
task, and then can be accessed from templates using
e.g. @turk_params['turkSubmitTo']. The hash is stored in the
Tasks table serialized as text and will be empty if the user
did not enter via Turk.

In order to convey the condition ID to the task, rather than using
the GetHIT web call as originally planned (which would have been complex
due to authentication requirements, and would have required tracking our
Turk credentials in the database), I just made it part of the entry point
URL as above.

Finally, if the ASSIGNMENT_ID_NOT_AVAILABLE assignmentId is observed,
the server needs to render a preview of how the task will appear, rather
than create a task. For this for the moment I use a simple ERB template
that renders a screenshot image and nothing else.

This should be all the functionality needed for Turk but further testing
on the live Heroku and Turk sites is necessary.